prepareData function with serverObject
In order to be less restrective prepareData receives as parameter the subset and the full server object. This allow us to manage extra data set by the server. In my case I'm sending a total row with the column sums, and other data, in order to add a new total row to the table.